Rename component #JC19-395
