+++ apps/angular/src/app/features/anime/anime-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { TableViewComponent } from './table-view.component';
+import { AnimeComponent } from './anime.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: TableViewComponent,
+    component: AnimeComponent,
     children: [],
   },
 ];
