+++ apps/angular/src/app/features/anime/anime.component.html
@@ -1,0 +1,7 @@
+<header class="header">
+  <h1 class="mat-headline">AnimeCatalog</h1>
+</header>
+
+<main class="content">
+  <anime-table-view></anime-table-view>
+</main>
