+++ apps/angular/src/app/features/anime/anime.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+/** Main page component. */
+@Component({
+  selector: 'anime',
+  templateUrl: './anime.component.html',
+  styleUrls: [],
+})
+export class AnimeComponent {}
