+++ apps/angular/src/app/features/anime/anime.module.ts
@@ -3,12 +3,13 @@ import { CommonModule } from '@angular/common';
 import { MatTableModule } from '@angular/material/table';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
-import { TableViewComponent } from './table-view.component';
-import { TableViewRoutingModule } from './table-view-routing.module';
+import { AnimeComponent } from './anime.component';
+import { TableViewRoutingModule } from './anime-routing.module';
+import { TableViewComponent } from './table-view/table-view.component';
 
 /** Table view module. */
 @NgModule({
-  declarations: [TableViewComponent],
+  declarations: [AnimeComponent, TableViewComponent],
   imports: [
     CommonModule,
     MatTableModule,
@@ -16,4 +17,4 @@ import { TableViewRoutingModule } from './table-view-routing.module';
     MatProgressSpinnerModule,
   ],
 })
-export class TableViewModule { }
+export class AnimeModule { }
