+++ apps/angular/src/app/features/anime/table-view/table-view.component.css
@@ -1,5 +1,6 @@
 .table-view {
   display: flex;
+  flex-direction: column;
   justify-content: center;
   align-items: center;
 }
