+++ apps/angular/src/app/features/anime/table-view/table-view.component.html
@@ -1,5 +1,7 @@
 <section class="table-view">
-  <table *ngIf="animeList$ | async as anime; else load" mat-table [dataSource]="anime" class="mat-elevation-z8 demo-table">
+  <h2 id="anime-catalog" class="mat-title">Anime Catalog</h2>
+  <table *ngIf="animeList$ | async as anime; else load" mat-table [dataSource]="anime"
+    class="mat-elevation-z8 demo-table" aria-label="anime" aria-describedby="catalog-anime">
     <ng-container matColumnDef="image">
       <th mat-header-cell *matHeaderCellDef>Image</th>
       <td mat-cell *matCellDef="let element"><img class="table-image" src="{{element.image}}"></td>
