+++ apps/angular/src/index.html
@@ -13,7 +13,7 @@
 </head>
 
 <body>
-  <anime-root></anime-root>
+  <app-root></app-root>
 </body>
 
 </html>
